// gamesimon/src/utilities/enums.tsx
export enum SCREENS {
    HomePage = "HomePage",
    ScoresPage = "ScoresPage",
}

export enum Colors {
    RED = 1,
    BLUE = 2,
    GREEN = 3,
    YELLOW = 4
}
export enum SoundByColors {
    RED = 'soundRed.mp3',
    BLUE = 'soundBlue.mp3',
    GREEN = 'soundGreen.mp3',
    YELLOW = 'soundYellow.mp3'
}